scraper/requirements.txt
PyVirtualDisplay==3.0
selenium==4.23.0
requests==2.32.3
lxml==5.2.2
pymongo==4.8.0
halo==0.0.31
PyYAML==6.0.1
